newpwork_backend_new/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-allauth==65.15.1
Pillow==11.3.0
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
sqlparse==0.5.3
stripe==11.5.0
tzdata==2025.2
requests==2.32.3
cryptography==46.0.6
